feat: add ApplicationDetailsSheet component for displaying application details; integrate confirmation dialog in AcceptOrRejectApplicationForm for decision handling
